+++ apps/game_services_web/lib/game_services_web/controllers/registration_controller.ex
@@ -6,7 +6,10 @@ defmodule GameServicesWeb.RegistrationController do
   end
 
 
-  def register(conn, _params) do
+  def register(conn, %{"register" => user_params}) do
+    IO.puts inspect user_params
+  {:ok, user} = UserManager.register_new_user(Map.get(user_params, "username"), Map.get(user_params, "email"), Map.get(user_params, "password"))
+  IO.puts inspect user
     render(conn, "index.html")
   end