+++ apps/user_manager/test/user_manager_test.exs
@@ -1,0 +1,5 @@
+defmodule UserManagerTest do
+  use ExUnit.Case
+  doctest UserManager
+
+end