+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 * integrate dialyzer
 * integrate dockerhub for base images to speed up deployment
 * integrate circle-ci
+* user registration workflow